id: 修正v2info回傳point為小數點時，處理數據會錯誤的問題

See merge request hm/hm-video-app!1006

+++ packages/shared/lib/utils/intl_amount.dart
@@ -4,7 +4,7 @@ import 'package:game/models/third_login_api_response_with_data.dart';
 import 'package:get_storage/get_storage.dart';
 import 'package:intl/intl.dart';
 
-String intlAmount(int amount, String currency) {
+String intlAmount(double amount, String currency) {
   var getLocale = GetStorage('locale').read('locale');
   return '${amount > 0 ? NumberFormat.currency(
       locale: getLocale ?? 'zh-TW',
